Added One way walls and Doors with switches
One way walls are regular solid objects however they start inactive and are only made active when an object triggers its trigger box, meaning something can jump through it one side and not the other.

Doors and switches are both children of the same object where the door relating to the switch is turn off or on when a switch is hit.

Also refined the jump to not need jump reset boxes.