Created the dash ability
I created the code for the dash ability, as well as tying the heavy attack to the heavy attack ability. This seperates the heavy attack from the regular attack by making it exclusive to the pink character. Ontop of this, I have combined the 2 types of blocks to now be the same thing, meaning you can choose to destroy it or knock it around.